requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
openai==1.3.6
Werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
Flask-Migrate==4.0.5
python-dotenv==1.0.1
pyyaml


google-generativeai==0.8.4